downstream/requirements.txt
# AncST downstream analysis dependencies
#
# IMPORTANT: Requires Python 3.11 (recommended) or 3.9/3.10
# (Python 3.12+ is NOT compatible - scipy and other packages require <3.12)
#
# Installation with conda (recommended):
#   conda create -n ancst_downstream python=3.11
#   conda activate ancst_downstream
#   pip install -r requirements.txt
#
# Installation with venv:
#   python3.11 -m venv ancst_downstream
#   source ancst_downstream/bin/activate
#   pip install -r requirements.txt
#
# Installation with pyenv:
#   pyenv install 3.11.10
#   pyenv shell 3.11.10
#   python3 -m venv ancst_downstream
#   source ancst_downstream/bin/activate
#   pip install -r requirements.txt
#
# Note: pandas can be commented out if overhead is a concern,
#       but make_ortho_tables.py and plot_ortho_stats.py will not work

numpy~=1.25.0
scipy~=1.10.0
biopython~=1.81
matplotlib~=3.7.0
networkx>=3.0
markov-clustering==0.0.6.dev0
dbscan1d==0.2.3
ete3==3.1.3
tralda>=0.2.0
pygenomeviz==1.3.0
pandas~=2.0.3
